feat: add merge facade api can configure whether to pop up diglog
