+++ packages/sheets/src/commands/commands/add-worksheet-merge.command.ts
@@ -14,7 +14,13 @@
  * limitations under the License.
  */
 
-import type { IAccessor, ICellData, ICommand, IMutationInfo, Injector, IRange, Nullable, Workbook, Worksheet } from '@univerjs/core';
+import type { IAccessor, ICellData, ICommand, IMutationInfo, IRange, Nullable, Workbook, Worksheet } from '@univerjs/core';
+import type {
+    IAddWorksheetMergeMutationParams,
+    IRemoveWorksheetMergeMutationParams,
+} from '../../basics/interfaces/mutation-interface';
+
+import type { ISetRangeValuesMutationParams } from '../mutations/set-range-values.mutation';
 import {
     CommandType,
     Dimension,
@@ -22,15 +28,9 @@ import {
     IUndoRedoService,
     IUniverInstanceService,
     ObjectMatrix,
-    Rectangle,
     sequenceExecute,
     UniverInstanceType,
 } from '@univerjs/core';
-
-import type {
-    IAddWorksheetMergeMutationParams,
-    IRemoveWorksheetMergeMutationParams,
-} from '../../basics/interfaces/mutation-interface';
 import { getAddMergeMutationRangeByType } from '../../controllers/merge-cell.controller';
 import { SheetsSelectionsService } from '../../services/selections/selection-manager.service';
 import { AddMergeUndoMutationFactory, AddWorksheetMergeMutation } from '../mutations/add-worksheet-merge.mutation';
@@ -38,9 +38,7 @@ import {
     RemoveMergeUndoMutationFactory,
     RemoveWorksheetMergeMutation,
 } from '../mutations/remove-worksheet-merge.mutation';
-import type { ISetRangeValuesMutationParams } from '../mutations/set-range-values.mutation';
 import { SetRangeValuesMutation, SetRangeValuesUndoMutationFactory } from '../mutations/set-range-values.mutation';
-import { getSheetCommandTarget } from './utils/target-util';
 
 export interface IAddMergeCommandParams {
     value?: Dimension.ROWS | Dimension.COLUMNS;
@@ -265,24 +263,3 @@ export const AddWorksheetMergeHorizontalCommand: ICommand = {
     },
 };
 
-export async function addMergeCellsUtil(injector: Injector, unitId: string, subUnitId: string, ranges: IRange[]) {
-    const univerInstanceService = injector.get(IUniverInstanceService);
-    const target = getSheetCommandTarget(univerInstanceService, { unitId, subUnitId });
-    if (!target) return;
-    const { worksheet } = target;
-    const mergeData = worksheet.getMergeData();
-    const overlap = mergeData.some((mergeRange) => {
-        return ranges.some((range) => {
-            return Rectangle.intersects(range, mergeRange);
-        });
-    });
-    if (overlap) {
-        throw new Error('The ranges to be merged overlap with the existing merged cells');
-    }
-    const commandService = injector.get(ICommandService);
-    await commandService.executeCommand(AddWorksheetMergeCommand.id, {
-        unitId,
-        subUnitId,
-        selections: ranges,
-    });
-}
